updated leave

+++ src/components/Admin/RejectedLeave.js
@@ -64,8 +64,8 @@ function RejectedLeave() {
                   <th>{item?.fullName}</th>
                   <th>{item?.employee_id?.role?.name}</th>
                   <th>{item?.leaveType}</th>
-                  <th>{new Date(item.startDate).toLocaleDateString('en-GB',{day:'numeric',month:'long',year:'numeric'})}</th>
-                  <th>{new Date(item.endDate).toLocaleDateString('en-GB',{day:'numeric',month:'long',year:'numeric'})}</th>
+                  <th>{new Date(item.startDate).toLocaleDateString('en-GB',{day:'numeric',month:'short',year:'numeric'})}</th>
+                  <th>{new Date(item.endDate).toLocaleDateString('en-GB',{day:'numeric',month:'short',year:'numeric'})}</th>
                   <th>{CalculateDays(item.startDate,item.endDate)} days</th>
                   <th>
                     <button className={`status-button1 ${getStatusColorClass(item?.status)}`}>
